Move form utilities to their own package
